feat: scripts for testnet contract deployment and holesky test
