+++ build.rs
@@ -5,6 +5,5 @@ fn main() {
         "./contracts/lib/forge-std",
         "./contracts",
     ];
-
     blueprint_build_utils::build_contracts(contract_dirs);
 }
